+++ packages/theme/index.ts
@@ -1,3 +1,5 @@
 import * as plugins from "./plugins";
 
-export const aristobyteui = Object.values(plugins);
+export const aristobyteui = () => Object.values(plugins);
+export * from "./plugins";
+export default aristobyteui;