+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aristobyte-ui/cli
 
+## 1.0.60
+
+### Patch Changes
+
+- feat: create theme package for aristobyteui plugins
+
+- Updated dependencies []:
+  - @aristobyte-ui/eslint-config@1.0.60
+  - @aristobyte-ui/typescript-config@1.0.60
+
 ## 1.0.59
 
 ### Patch Changes