+++ packages/message-box/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aristobyte-ui/message-box
 
+## 1.0.60
+
+### Patch Changes
+
+- feat: create theme package for aristobyteui plugins
+
+- Updated dependencies []:
+  - @aristobyte-ui/utils@1.0.60
+
 ## 1.0.59
 
 ### Patch Changes