+++ packages/presets/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @aristobyte-ui/presets
 
+## 1.0.60
+
+### Patch Changes
+
+- feat: create theme package for aristobyteui plugins
+
+- Updated dependencies []:
+  - @aristobyte-ui/anchor@1.0.60
+  - @aristobyte-ui/button@1.0.60
+  - @aristobyte-ui/utils@1.0.60
+
 ## 1.0.59
 
 ### Patch Changes